+++ src/config.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'languages' => ['markup', 'javascript', 'css', 'twig'],
+    'themes' => ['*'],
+    'plugins' => ['line-numbers'],
+];